moves scorecard component into scorekeeper page
